wip: Use browser resolve
