+++ src/lib/crypto.ts
@@ -1,26 +1,4 @@
-import nodeCrypto from "crypto";
-
-// IE 11
-declare global {
-  interface Window {
-    msCrypto?: Crypto;
-  }
-
-  interface Crypto {
-    webkitSubtle?: SubtleCrypto;
-  }
-}
-
-const crypto =
-  (typeof window !== "undefined" &&
-    (window as Window) &&
-    (window.crypto || window.msCrypto)) ||
-  (nodeCrypto.webcrypto as unknown as Crypto);
-const subtle: SubtleCrypto = crypto.subtle || crypto.webkitSubtle;
-
-if (subtle === undefined) {
-  throw new Error("crypto and/or subtle api unavailable");
-}
+import { crypto, subtle } from "./crypto_subtle";
 
 export { crypto, subtle };
